refactor: update import statements and improve layout settings in number partitioning scripts
