+++ examples/number_partitioning/npp_paths.py
